midenc-compile: separate target synthesis from test helpers
